fix(error-tracking): break circular import between query runners and api package
build_skills.py started crashing on master with "cannot import name
'ErrorTrackingQueryRunner' from partially initialized module" after the
lazy DRF router change (#60849). The cycle was always there, but eager
route registration during django.setup() used to fully initialize the
api package before anything else imported the runners, masking it. The
skills build imports the runners in a cold interpreter, where the cycle
is live: runner -> api.issues triggers api/__init__.py -> api/query.py
-> back into the runner.

The runner -> api edge was also backwards architecturally: the shared
serializers (ErrorTrackingIssuePreviewSerializer was not even used
inside the api package) now live in issue_serializers.py outside api/,
so the query layer no longer imports the viewset layer.

Add a cold-interpreter regression test mirroring how build_skills hits
the imports — pytest alone can't catch this because the modules are
already cached by the time tests run.
